requirements.txt
# Core Dependencies
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Google Gemini API
google-generativeai==0.3.2
google-cloud-vision==3.5.0  # Optional: if you want to use Vision API too

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3  # For PDF processing

# Data Processing
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# Text Processing & Normalization
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
langdetect==1.0.9
indic-transliteration==2.3.46  # For Indian language transliteration
unidecode==1.3.7  # For basic transliteration

# Utilities
requests==2.31.0
pydantic==2.5.2  # For data validation
marshmallow==3.20.1  # For serialization
loguru==0.7.2  # Better logging

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.0  # Code formatter
flake8==6.1.0  # Linter
ipython==8.18.1  # For debugging
